Driftsendepunkt for å slette filter * I førsteomgang tillat slett kun hvis man selv har opprettet filteret
